refactor(#3037): relocate kanban view DTOs below service layer
버킷4(도메인 타입 server/dto 오배치) 슬라이스. 서비스 전용 응답 타입
`KanbanCardView`/`ListCardsResponse`와 `From<KanbanCardRecord>` impl을
`server/dto/kanban.rs`에서 `services/kanban_cards.rs`로 하향 이동한다.

두 타입은 `KanbanService::list_cards`가 유일한 소비처이고 server route는
참조하지 않는다(전수 grep 확인). 의존성도 `crate::db::kanban_cards`와
`crate::utils::github_links`로 모두 service 계층 이하라 순수 relocation이다.

- src/services/kanban_cards.rs:4 의 `use crate::server::dto::kanban::*`
  backflow import 제거 → service→server 역참조 1건 소멸
- `services::kanban`이 `kanban_cards::*`를 re-export하므로
  route의 `crate::services::kanban::ListCardsInput` 등 호출부는 무변경
- service_server_backflow baseline: kanban_cards.rs 엔트리 제거,
  total_count 44→43

동작 보존: 직렬화 형태/필드/From 변환 로직 동일, import 경로만 변경.
